quiz: allow dropping samples without failing
It is often useful to drop samples for which the test is ill-defined.
This should not result in the overall test crashing, for it is not the
same as a failure. However it is also doesn't exercise the code under
test. Therefore it should be reported that the sample was simply
dropped and replaced.